refactor(core): separate reflective injector from Injector interface
