+++ modules/angular2/examples/core/di/ts/forward_ref/forward_ref.ts
@@ -1,4 +1,10 @@
-import {Inject, Injector, forwardRef, resolveForwardRef, ForwardRefFn} from 'angular2/core';
+import {
+  Inject,
+  ReflectiveInjector,
+  forwardRef,
+  resolveForwardRef,
+  ForwardRefFn
+} from 'angular2/core';
 
 // #docregion forward_ref_fn
 var ref = forwardRef(() => Lock);
@@ -13,7 +19,7 @@ class Door {
 // Only at this point Lock is defined.
 class Lock {}
 
-var injector = Injector.resolveAndCreate([Door, Lock]);
+var injector = ReflectiveInjector.resolveAndCreate([Door, Lock]);
 var door = injector.get(Door);
 expect(door instanceof Door).toBe(true);
 expect(door.lock instanceof Lock).toBe(true);
