+++ modules/angular2/src/compiler/output/interpretive_view.ts
@@ -19,7 +19,7 @@ class _InterpretiveAppView extends AppView<any> implements DynamicInstance {
     super(args[0], args[1], args[2], args[3], args[4], args[5], args[6], args[7], args[8], args[9],
           args[10]);
   }
-  createInternal(rootSelector: string): void {
+  createInternal(rootSelector: string | any): void {
     var m = this.methods.get('createInternal');
     if (isPresent(m)) {
       m(rootSelector);
