+++ modules/angular2/src/core/di/reflective_exceptions.ts
@@ -2,7 +2,7 @@ import {ListWrapper} from 'angular2/src/facade/collection';
 import {stringify, isBlank} from 'angular2/src/facade/lang';
 import {BaseException, WrappedException, unimplemented} from 'angular2/src/facade/exceptions';
 import {Key} from './key';
-import {Injector} from './injector';
+import {ReflectiveInjector} from './reflective_injector';
 
 function findFirstClosedCycle(keys: any[]): any[] {
   var res = [];
@@ -39,12 +39,12 @@ export class AbstractProviderError extends BaseException {
   keys: Key[];
 
   /** @internal */
-  injectors: Injector[];
+  injectors: ReflectiveInjector[];
 
   /** @internal */
   constructResolvingMessage: Function;
 
-  constructor(injector: Injector, key: Key, constructResolvingMessage: Function) {
+  constructor(injector: ReflectiveInjector, key: Key, constructResolvingMessage: Function) {
     super("DI Exception");
     this.keys = [key];
     this.injectors = [injector];
@@ -52,7 +52,7 @@ export class AbstractProviderError extends BaseException {
     this.message = this.constructResolvingMessage(this.keys);
   }
 
-  addKey(injector: Injector, key: Key): void {
+  addKey(injector: ReflectiveInjector, key: Key): void {
     this.injectors.push(injector);
     this.keys.push(key);
     this.message = this.constructResolvingMessage(this.keys);
@@ -76,7 +76,7 @@ export class AbstractProviderError extends BaseException {
  * ```
  */
 export class NoProviderError extends AbstractProviderError {
-  constructor(injector: Injector, key: Key) {
+  constructor(injector: ReflectiveInjector, key: Key) {
     super(injector, key, function(keys: any[]) {
       var first = stringify(ListWrapper.first(keys).token);
       return `No provider for ${first}!${constructResolvingPath(keys)}`;
@@ -101,7 +101,7 @@ export class NoProviderError extends AbstractProviderError {
  * Retrieving `A` or `B` throws a `CyclicDependencyError` as the graph above cannot be constructed.
  */
 export class CyclicDependencyError extends AbstractProviderError {
-  constructor(injector: Injector, key: Key) {
+  constructor(injector: ReflectiveInjector, key: Key) {
     super(injector, key, function(keys: any[]) {
       return `Cannot instantiate cyclic dependency!${constructResolvingPath(keys)}`;
     });
@@ -139,15 +139,15 @@ export class InstantiationError extends WrappedException {
   keys: Key[];
 
   /** @internal */
-  injectors: Injector[];
+  injectors: ReflectiveInjector[];
 
-  constructor(injector: Injector, originalException, originalStack, key: Key) {
+  constructor(injector: ReflectiveInjector, originalException, originalStack, key: Key) {
     super("DI Exception", originalException, originalStack, null);
     this.keys = [key];
     this.injectors = [injector];
   }
 
-  addKey(injector: Injector, key: Key): void {
+  addKey(injector: ReflectiveInjector, key: Key): void {
     this.injectors.push(injector);
     this.keys.push(key);
   }
