+++ modules/angular2/src/core/linker/element_injector.ts
@@ -1,22 +1,19 @@
-import {isBlank, stringify} from 'angular2/src/facade/lang';
-import {Injector, UNDEFINED} from 'angular2/src/core/di/injector';
+import {isBlank, stringify, CONST_EXPR} from 'angular2/src/facade/lang';
+import {Injector, THROW_IF_NOT_FOUND} from 'angular2/src/core/di/injector';
 import {AppView} from './view';
+
+const _UNDEFINED = CONST_EXPR(new Object());
 
 export class ElementInjector extends Injector {
   constructor(private _view: AppView<any>, private _nodeIndex: number) { super(); }
 
-  get(token: any): any {
-    var result = this._view.injectorGet(token, this._nodeIndex, UNDEFINED);
-    if (result === UNDEFINED) {
-      result = this._view.parentInjector.get(token);
+  get(token: any, notFoundValue: any = THROW_IF_NOT_FOUND): any {
+    var result = _UNDEFINED;
+    if (result === _UNDEFINED) {
+      result = this._view.injectorGet(token, this._nodeIndex, _UNDEFINED);
     }
-    return result;
-  }
-
-  getOptional(token: any): any {
-    var result = this._view.injectorGet(token, this._nodeIndex, UNDEFINED);
-    if (result === UNDEFINED) {
-      result = this._view.parentInjector.getOptional(token);
+    if (result === _UNDEFINED) {
+      result = this._view.parentInjector.get(token, notFoundValue);
     }
     return result;
   }
