+++ modules/angular2/src/router/directives/router_outlet.ts
@@ -8,9 +8,8 @@ import {
   DynamicComponentLoader,
   ComponentRef,
   ElementRef,
-  Injector,
   provide,
-  Dependency,
+  ReflectiveInjector,
   OnDestroy,
   Output
 } from 'angular2/core';
@@ -60,7 +59,7 @@ export class RouterOutlet implements OnDestroy {
     var componentType = nextInstruction.componentType;
     var childRouter = this._parentRouter.childRouter(componentType);
 
-    var providers = Injector.resolve([
+    var providers = ReflectiveInjector.resolve([
       provide(RouteData, {useValue: nextInstruction.routeData}),
       provide(RouteParams, {useValue: new RouteParams(nextInstruction.params)}),
       provide(routerMod.Router, {useValue: childRouter})
