+++ modules/angular2/test/http/backends/jsonp_backend_spec.ts
@@ -19,7 +19,7 @@ import {
   JSONPBackend,
   JSONPBackend_
 } from 'angular2/src/http/backends/jsonp_backend';
-import {provide, Injector} from 'angular2/core';
+import {provide, Injector, ReflectiveInjector} from 'angular2/core';
 import {isPresent, StringWrapper} from 'angular2/src/facade/lang';
 import {TimerWrapper} from 'angular2/src/facade/async';
 import {Request} from 'angular2/src/http/static_request';
@@ -71,7 +71,7 @@ export function main() {
     let sampleRequest: Request;
 
     beforeEach(() => {
-      let injector = Injector.resolveAndCreate([
+      let injector = ReflectiveInjector.resolveAndCreate([
         provide(ResponseOptions, {useClass: BaseResponseOptions}),
         provide(BrowserJsonp, {useClass: MockBrowserJsonp}),
         provide(JSONPBackend, {useClass: JSONPBackend_})
