+++ modules/angular2/test/http/backends/mock_backend_spec.ts
@@ -14,7 +14,7 @@ import {
 import {ObservableWrapper} from 'angular2/src/facade/async';
 import {BrowserXhr} from 'angular2/src/http/backends/browser_xhr';
 import {MockConnection, MockBackend} from 'angular2/src/http/backends/mock_backend';
-import {provide, Injector} from 'angular2/core';
+import {provide, Injector, ReflectiveInjector} from 'angular2/core';
 import {Request} from 'angular2/src/http/static_request';
 import {Response} from 'angular2/src/http/static_response';
 import {Headers} from 'angular2/src/http/headers';
@@ -34,7 +34,7 @@ export function main() {
     var sampleResponse2: Response;
 
     beforeEach(() => {
-      var injector = Injector.resolveAndCreate(
+      var injector = ReflectiveInjector.resolveAndCreate(
           [provide(ResponseOptions, {useClass: BaseResponseOptions}), MockBackend]);
       backend = injector.get(MockBackend);
       var base = new BaseRequestOptions();
