+++ modules/angular2/test/platform/browser/tools/spies.dart
@@ -1,5 +1,5 @@
 import 'package:angular2/testing_internal.dart' show SpyObject;
-import 'package:angular2/core.dart' show Injector, bind;
+import 'package:angular2/core.dart' show Injector, ReflectiveInjector, bind;
 import 'package:angular2/src/core/application_ref.dart' show ApplicationRef;
 import 'package:angular2/src/core/linker/component_factory.dart'
     show ComponentRef;
@@ -15,7 +15,7 @@ class SpyComponentRef extends SpyObject implements ComponentRef {
   Injector injector;
 
   SpyComponentRef() {
-    this.injector = Injector
+    this.injector = ReflectiveInjector
         .resolveAndCreate([bind(ApplicationRef).toClass(SpyApplicationRef)]);
   }
 }
