+++ modules/angular2/test/public_api_spec.ts
@@ -169,7 +169,7 @@ var NG_CORE = [
   'PLATFORM_PIPES',
   'DebugNode',
   'DebugElement',
-  'Dependency',
+  'ReflectiveDependency',
   'DependencyMetadata',
   'Directive',
   'DirectiveMetadata',
@@ -194,6 +194,7 @@ var NG_CORE = [
   'Injectable',
   'InjectableMetadata',
   'Injector',
+  'ReflectiveInjector',
   'InstantiationError',
   'InvalidProviderError',
   'IterableDiffers',
@@ -219,9 +220,9 @@ var NG_CORE = [
   'Renderer',
   'RootRenderer',
   'RenderComponentType',
-  'ResolvedBinding:dart',
-  'ResolvedProvider:dart',
-  'ResolvedFactory',
+  'ResolvedReflectiveBinding:dart',
+  'ResolvedReflectiveProvider:dart',
+  'ResolvedReflectiveFactory',
   'Self',
   'SelfMetadata',
   'SkipSelf',
