+++ modules/angular2/test/router/location/location_spec.ts
@@ -12,7 +12,7 @@ import {
   SpyObject
 } from 'angular2/testing_internal';
 
-import {Injector, provide} from 'angular2/core';
+import {Injector, provide, ReflectiveInjector} from 'angular2/core';
 import {CONST_EXPR} from 'angular2/src/facade/lang';
 
 import {Location} from 'angular2/src/router/location/location';
@@ -27,7 +27,7 @@ export function main() {
     function makeLocation(baseHref: string = '/my/app', provider: any = CONST_EXPR([])): Location {
       locationStrategy = new MockLocationStrategy();
       locationStrategy.internalBaseHref = baseHref;
-      let injector = Injector.resolveAndCreate(
+      let injector = ReflectiveInjector.resolveAndCreate(
           [Location, provide(LocationStrategy, {useValue: locationStrategy}), provider]);
       return location = injector.get(Location);
     }
