+++ modules/benchpress/src/runner.ts
@@ -1,4 +1,4 @@
-import {Injector, bind, provide, Provider} from 'angular2/src/core/di';
+import {Injector, bind, provide, Provider, ReflectiveInjector} from 'angular2/src/core/di';
 import {isPresent, isBlank} from 'angular2/src/facade/lang';
 import {PromiseWrapper} from 'angular2/src/facade/async';
 
@@ -52,7 +52,7 @@ export class Runner {
       sampleBindings.push(bindings);
     }
 
-    var inj = Injector.resolveAndCreate(sampleBindings);
+    var inj = ReflectiveInjector.resolveAndCreate(sampleBindings);
     var adapter = inj.get(WebDriverAdapter);
 
     return PromiseWrapper
@@ -66,7 +66,7 @@ export class Runner {
           // Only WebDriverAdapter is reused.
           // TODO vsavkin consider changing it when toAsyncFactory is added back or when child
           // injectors are handled better.
-          var injector = Injector.resolveAndCreate([
+          var injector = ReflectiveInjector.resolveAndCreate([
             sampleBindings,
             bind(Options.CAPABILITIES).toValue(capabilities),
             bind(Options.USER_AGENT).toValue(userAgent),
