+++ modules/benchpress/test/reporter/json_file_reporter_spec.ts
@@ -17,7 +17,7 @@ import {PromiseWrapper} from 'angular2/src/facade/async';
 import {
   bind,
   provide,
-  Injector,
+  ReflectiveInjector,
   SampleDescription,
   MeasureValues,
   Options
@@ -43,7 +43,7 @@ export function main() {
               return PromiseWrapper.resolve(null);
             })
       ];
-      return Injector.resolveAndCreate(bindings).get(JsonFileReporter);
+      return ReflectiveInjector.resolveAndCreate(bindings).get(JsonFileReporter);
     }
 
     it('should write all data into a file', inject([AsyncTestCompleter], (async) => {
