+++ modules/benchpress/test/reporter/multi_reporter_spec.ts
@@ -14,14 +14,21 @@ import {
 import {PromiseWrapper} from 'angular2/src/facade/async';
 import {DateWrapper} from 'angular2/src/facade/lang';
 
-import {Reporter, MultiReporter, bind, provide, Injector, MeasureValues} from 'benchpress/common';
+import {
+  Reporter,
+  MultiReporter,
+  bind,
+  provide,
+  ReflectiveInjector,
+  MeasureValues
+} from 'benchpress/common';
 
 export function main() {
   function createReporters(ids: any[]) {
-    var r = Injector.resolveAndCreate([
-                      ids.map(id => provide(id, {useValue: new MockReporter(id)})),
-                      MultiReporter.createBindings(ids)
-                    ])
+    var r = ReflectiveInjector.resolveAndCreate([
+                                ids.map(id => provide(id, {useValue: new MockReporter(id)})),
+                                MultiReporter.createBindings(ids)
+                              ])
                 .get(MultiReporter);
     return PromiseWrapper.resolve(r);
   }
