+++ modules/benchpress/test/web_driver_extension_spec.ts
@@ -14,16 +14,16 @@ import {
 import {isPresent, StringWrapper} from 'angular2/src/facade/lang';
 import {PromiseWrapper} from 'angular2/src/facade/async';
 
-import {WebDriverExtension, bind, provide, Injector, Options} from 'benchpress/common';
+import {WebDriverExtension, bind, provide, ReflectiveInjector, Options} from 'benchpress/common';
 
 export function main() {
   function createExtension(ids: any[], caps) {
     return PromiseWrapper.wrap(() => {
-      return Injector.resolveAndCreate([
-                       ids.map(id => provide(id, {useValue: new MockExtension(id)})),
-                       bind(Options.CAPABILITIES).toValue(caps),
-                       WebDriverExtension.bindTo(ids)
-                     ])
+      return ReflectiveInjector.resolveAndCreate([
+                                 ids.map(id => provide(id, {useValue: new MockExtension(id)})),
+                                 bind(Options.CAPABILITIES).toValue(caps),
+                                 WebDriverExtension.bindTo(ids)
+                               ])
           .get(WebDriverExtension);
     });
   }
